# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MediAsk API",
    "version": "1.0.0",
    "description": "智能医疗辅助问诊系统 - 按模块组织"
  },
  "servers": [
    {
      "url": "http://localhost:8989",
      "description": "开发环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/appointments/doctor/{doctorId}": {
      "get": {
        "tags": [
          "appointment"
        ],
        "summary": "查询医生预约列表（按日期）",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "查询日期"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "getApiV1AppointmentsDoctorByDoctorId",
        "description": "查询医生在指定日期的预约列表（doctor 权限）。"
      }
    },
...
    "/api/v1/doctors": {
      "post": {
        "tags": [
          "doctor"
        ],
        "summary": "创建医生档案",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDoctorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "postApiV1Doctors"
      },
      "get": {
        "tags": [
          "doctor"
        ],
        "summary": "分页查询医生档案",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "用户ID"
          },
          {
            "name": "doctorCode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "医生编码"
          },
          {
            "name": "hospitalId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "医院ID"
          },
          {
            "name": "departmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "科室ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "状态 0-停用 1-启用"
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "关键字（医生编码/职称/执业证号）"
          },
          {
            "name": "pageNum",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "页码",
            "example": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "每页条数",
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultResponse"
                }
              }
            }
          }
        },
        "operationId": "getApiV1Doctors"
      }
    },
    "/api/v1/doctors/{doctorId}": {
      "get": {
        "tags": [
          "doctor"
        ],
        "summary": "医生详情",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "医生ID"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "getApiV1DoctorsByDoctorId"
      },
      "put": {
        "tags": [
          "doctor"
        ],
        "summary": "更新医生档案",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "医生ID"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDoctorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "putApiV1DoctorsByDoctorId"
      }
    },
    "/api/v1/doctors/{doctorId}/status": {
      "put": {
        "tags": [
          "doctor"
        ],
        "summary": "启停医生档案",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "医生ID"
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "enum": [
                0,
                1
              ]
            },
            "description": "状态：0=停用，1=启用"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "putApiV1DoctorsByDoctorIdStatus"
      }
    },
...
    "/api/v1/schedules/doctor/{doctorId}": {
      "get": {
        "tags": [
          "schedule"
        ],
        "summary": "医生排班列表",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "doctorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseWrapper"
                }
              }
            }
          }
        },
        "operationId": "getApiV1SchedulesDoctorByDoctorId"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateDoctorRequest": {
        "type": "object",
        "required": [
          "userId",
          "hospitalId",
          "departmentId",
          "doctorCode"
        ],
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "hospitalId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorCode": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "introduction": {
            "type": "string"
          },
          "consultationFee": {
            "type": "number",
            "format": "double"
          },
          "licenseNumber": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PageResultResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "format": "int64",
                "description": "总记录数"
              },
              "pageNum": {
                "type": "integer",
                "format": "int32",
                "description": "页码"
              },
              "pageSize": {
                "type": "integer",
                "format": "int32",
                "description": "每页数量"
              },
              "list": {
                "type": "array",
                "description": "数据列表"
              }
            }
          }
        }
      },
...
      "ResponseWrapper": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "description": "状态码，0表示成功"
          },
          "msg": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "type": "object",
            "description": "响应数据"
          },
          "traceId": {
            "type": "string",
            "description": "链路追踪ID"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "响应时间戳（毫秒）"
          }
        },
        "example": {
          "code": 0,
          "msg": "success",
          "data": {}
        }
      },
...
      "UpdateDoctorRequest": {
        "type": "object",
        "properties": {
          "hospitalId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "doctorCode": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "introduction": {
            "type": "string"
          },
          "consultationFee": {
            "type": "number",
            "format": "double"
          },
          "licenseNumber": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}